+++ gems/aws-sdk-core/lib/aws-sdk-sso/client.rb
@@ -696,7 +696,7 @@ module Aws::SSO
         tracer: tracer
       )
       context[:gem_name] = 'aws-sdk-core'
-      context[:gem_version] = '3.252.0'
+      context[:gem_version] = '3.253.0'
       Seahorse::Client::Request.new(handlers, context)
     end