+++ gems/aws-sdk-core/lib/aws-sdk-ssooidc/client.rb
@@ -1079,7 +1079,7 @@ module Aws::SSOOIDC
         tracer: tracer
       )
       context[:gem_name] = 'aws-sdk-core'
-      context[:gem_version] = '3.252.0'
+      context[:gem_version] = '3.253.0'
       Seahorse::Client::Request.new(handlers, context)
     end