+++ gems/aws-sdk-core/lib/aws-sdk-sts/client.rb
@@ -2723,7 +2723,7 @@ module Aws::STS
         tracer: tracer
       )
       context[:gem_name] = 'aws-sdk-core'
-      context[:gem_version] = '3.252.0'
+      context[:gem_version] = '3.253.0'
       Seahorse::Client::Request.new(handlers, context)
     end